Crate ft_matrix

Vector and Matrix Library

This library provides functionality for working with vectors and matrices. It includes utility functions for operations like calculating size, reshaping, and printing.

Re-exports

pub use vector::Vector;

pub use matrix::Matrix;

pub use scalar::Scalar;

Modules

linear_interpolation

The linear interpolation module provides a generic function implementation to apply an interpolation.

matrix

The matrix module provides a generic Matrix struct and related methods.

scalar

The scalar module provides a trait generic for scalars.

vector

The vector module provides a generic Vector struct and related methods.