Struct ft_matrix::matrix::Matrix

pub struct Matrix<K: Scalar> {
    pub data: Vec<Vec<K>>,
}

A struct representing a mathematical matrix.

The Matrix struct is generic over type K that implements the Scalar type, that which ensures that it supports basic arithmetic operations like addition, subtraction, multiplication, and division.

Fields
data: Vec<Vec<K>>
Implementations

impl<K: Scalar> Matrix<K>

pub fn new(data: Vec<Vec<K>>) -> Self

Creates a new Matrix<K> from a matrix of matrices (rows of the matrix).

Arguments

data - A 2D matrix of K values representing the rows of the matrix.

Example

use ft_matrix::Matrix;

let mat = Matrix::new(vec![vec![1.0, 2.0], vec![3.0, 4.0]]);

pub fn size(&self) -> (usize, usize)

Returns the size (i.e., number of rows and columns) of the matrix.

Returns

A tuple (rows, cols) representing the dimensions of the matrix.

Example

use ft_matrix::Matrix;

let mat = Matrix::new(vec![vec![1.0, 2.0], vec![3.0, 4.0]]);
assert_eq!(mat.size(), (2, 2));

pub fn is_square(&self) -> bool

Returns whether the matrix is square.

A matrix is square if the number of rows equals the number of columns.

Example

use ft_matrix::Matrix;

let mat = Matrix::new(vec![vec![1.0, 2.0], vec![3.0, 4.0]]);
assert!(mat.is_square());

pub fn print(&self)

Prints the contents of the matrix to the standard output.

The matrix will be printed row by row, with each row on a new line, like:

[1.0, 2.0]
[3.0, 4.0]

Example

use ft_matrix::Matrix;

let mat = Matrix::new(vec![vec![1.0, 2.0], vec![3.0, 4.0]]);
mat.print();

pub fn flatten(&self) -> Vector<K>

Reshapes the Martix into a Vector by flattening its rows.

Example

use ft_matrix::Matrix;

let mat = Matrix::new(vec![vec![1.0, 2.0], vec![3.0, 4.0]]);
let vec = mat.flatten();
assert_eq!(vec.size(), 4);

pub fn add(&mut self, v: &Matrix<K>)

Adds another Matrix<K> to the calling Matrix<K>.

Arguments

v - A reference to the other Matrix<K> to add.

Panics

This function will panic if the matrixes are not of the same size.

Example

use ft_matrix::Matrix;

let mut mat1 = Matrix::new(vec![
    vec![42.0, 4.2],
    vec![1.0, 2.0]
]);
let mat2 = Matrix::new(vec![
    vec![-42.0, 4.2],
    vec![1.0, 2.0]
]);

mat1.add(&mat2);

assert_eq!(mat1.data, vec![
    vec![0.0, 8.4],
    vec![2.0, 4.0]
]);

pub fn sub(&mut self, v: &Matrix<K>)

Substracts another Matrix<K> to the calling Matrix<K>.

Arguments

v - A reference to the other Matrix<K> to subtract.

Panics

This function will panic if the Matrixs are not of the same size.

Example

use ft_matrix::Matrix;

let mut mat1 = Matrix::new(vec![
    vec![42.0, 4.2],
    vec![1.0, 2.0]
]);
let mat2 = Matrix::new(vec![
    vec![1.0, 2.0],
    vec![-42.0, 4.2]
]);

mat1.sub(&mat2);

assert_eq!(mat1.data, vec![
    vec![42.0 - 1.0, 4.2 - 2.0],
    vec![1.0 - -42.0, 2.0 - 4.2]
]);

pub fn scl(&mut self, a: K)

Scales the calling Matrix<K> by a factor a.

Each element in the Matrix is multiplied by the scalar a.

Arguments

a - The scaling factor.

Example

use ft_matrix::Matrix;

let mut mat = Matrix::new(vec![
    vec![1.0, 2.0],
    vec![3.0, 4.0]
]);

mat.scl(2.0); // Scale by 2

assert_eq!(mat.data, vec![
    vec![2.0, 4.0],
    vec![6.0, 8.0]
]);